+++ tests/test_download_and_read.py
@@ -2,16 +2,18 @@
 import itertools
 import tempfile
 
+import pandas as pd
+
 import glodap
 
 
-def test_response():
+def test_download_and_read():
     with tempfile.TemporaryDirectory() as tdir:
         for region, version in itertools.product(
             glodap.regions, glodap.versions
         ):
-            print(region, version)
-            glodap.download(region=region, version=version, gpath=tdir)
+            df = glodap.read(region=region, version=version, gpath=tdir)
+            assert isinstance(df, pd.DataFrame)
 
 
-# test_response()
+# test_download_and_read()