+++ tests/test_regions.py
@@ -1,4 +1,6 @@
 # %%
+import pandas as pd
+
 import glodap
 
 
@@ -8,4 +10,18 @@ def test_region_names():
     assert len(regions_short) == len(glodap.regions)
 
 
+def test_regions_latest():
+    df = glodap.arctic()
+    assert isinstance(df, pd.DataFrame)
+    df = glodap.atlantic()
+    assert isinstance(df, pd.DataFrame)
+    df = glodap.indian()
+    assert isinstance(df, pd.DataFrame)
+    df = glodap.pacific()
+    assert isinstance(df, pd.DataFrame)
+    df = glodap.world()
+    assert isinstance(df, pd.DataFrame)
+
+
 # test_region_names()
+# test_regions_latest()