+++ changelog.md
@@ -1,5 +1,10 @@
 # glodap changelog
 
+## v0.4 (forthcoming)
+
+  * Added options to download and import the v2.2016 (original v2) and v3.2026 (original v3) datasets.
+  * Switched to [Pooch](https://github.com/fatiando/pooch) for download handling.
+
 ## v0.3 (26 May 2025)
 
   * Renaming columns for PyCO2SYS v2 removed (no longer necessary due to PyCO2SYS updates).